refactor: move referral code from auth to referral campaign module
