+++ packages/backend-modules/auth/graphql/resolvers/User.js
@@ -3,7 +3,6 @@ const { findAllUserSessions } = require('../../lib/Sessions')
 const { enabledFirstFactors } = require('../../lib/Users')
 const AccessToken = require('../../lib/AccessToken')
 const Consents = require('../../lib/Consents')
-const { generateReferralCode } = require('../../lib/ReferralCode')
 
 const DEFAULT_USER_ACCESS_ROLES = ['admin', 'supporter']
 const BASICS_USER_ACCESS_ROLES = [...DEFAULT_USER_ACCESS_ROLES, 'editor']
@@ -83,8 +82,5 @@ module.exports = {
         pgdb,
       })
     }
-  },
-  referralCode: async (user, _, { pgdb }) => {
-    return user.referralCode || (await generateReferralCode(user, pgdb))
   },
 }
