+++ packages/backend-modules/referral-campaigns/lib/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  referralHandler: require('./referralHandler'),
+  referralCode: require('./referralCode'),
+}
